Compass values are working with offsets
Still no PID but I fixed the whole 100.0 10.0 problem.. offsets can be
applied by changing holding regs

+++ PSoC 4 Compass Modbus Slave.cydsn/Modbus.c
@@ -18,7 +18,7 @@
 #define SlaveAddress 5 //Change this to whatever you want
 #define CHECK_BIT(var,pos) !!((var) & (1 << (pos)))
 
-unsigned int  holdingReg[50];                                                                                                                                                               
+uint16  holdingReg[50];                                                                                                                                                               
 unsigned char coils[50];
 unsigned char received[125];
 unsigned char response[125]; //Enough to return all holding-r's
